kotlin2cpg: move ast creation logic into traits (#3515)
